// src/openapi/model/api-mypage-project-application-to-add.ts
/* tslint:disable */
/* eslint-disable */

/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 *
 * @export
 * @interface ApiMypageProjectApplicationToAdd
 */
export interface ApiMypageProjectApplicationToAdd {
    /**
     *
     * @type {number}
     * @memberof ApiMypageProjectApplicationToAdd
     */
    projectId?: number;
    /**
     *
     * @type {number}
     * @memberof ApiMypageProjectApplicationToAdd
     */
    contractorId?: number;
    /**
     *
     * @type {number}
     * @memberof ApiMypageProjectApplicationToAdd
     */
    contractorPrice?: number;
    /**
     *
     * @type {boolean}
     * @memberof ApiMypageProjectApplicationToAdd
     */
    immediateReceivingFlag?: boolean;
    /**
     *
     * @type {boolean}
     * @memberof ApiMypageProjectApplicationToAdd
     */
    negotiationFlag?: boolean;
    /**
     *
     * @type {boolean}
     * @memberof ApiMypageProjectApplicationToAdd
     */
    questionFlag?: boolean;
    /**
     *
     * @type {string}
     * @memberof ApiMypageProjectApplicationToAdd
     */
    contractorComment?: string;
}